refactor: re-organize src/ into subdirectories to better resemble the BioDynaMo directory project-structure
