+++ src/agents/cart_cell.cc
@@ -19,10 +19,7 @@
  * for the compiler-research.org organization.
  */
 
-#include "cart_cell.h"
-#include "hyperparams.h"
-#include "tumor_cell.h"
-#include "utils_aux.h"
+#include "agents/cart_cell.h"
 #include "core/agent/agent.h"
 #include "core/agent/agent_pointer.h"
 #include "core/agent/new_agent_event.h"
@@ -41,6 +38,9 @@
 #include <cstdint>
 #include <stdexcept>
 #include <string>
+#include "agents/tumor_cell.h"
+#include "params/hyperparams.h"
+#include "utils/utils_aux.h"
 
 namespace bdm {
