+++ src/agents/cart_cell.h
@@ -22,7 +22,6 @@
 #ifndef CART_CELL_H_
 #define CART_CELL_H_
 
-#include "tumor_cell.h"
 #include "core/agent/agent.h"
 #include "core/agent/agent_pointer.h"
 #include "core/agent/cell.h"
@@ -32,8 +31,9 @@
 #include "core/diffusion/diffusion_grid.h"
 #include "core/interaction_force.h"
 #include "core/real_t.h"
-#include "core/substance_interactor.h"
 #include "core/util/random.h"
+#include "../interfaces/substance_interactor.h"
+#include "agents/tumor_cell.h"
 
 namespace bdm {
