+++ src/agents/tumor_cell.cc
@@ -19,9 +19,7 @@
  * for the compiler-research.org organization.
  */
 
-#include "tumor_cell.h"
-#include "hyperparams.h"
-#include "utils_aux.h"
+#include "agents/tumor_cell.h"
 #include "core/agent/agent.h"
 #include "core/agent/cell_division_event.h"
 #include "core/agent/new_agent_event.h"
@@ -39,6 +37,8 @@
 #include <cstdint>
 #include <stdexcept>
 #include <string>
+#include "params/hyperparams.h"
+#include "utils/utils_aux.h"
 
 namespace bdm {
