+++ src/diffusion/diffusion_thomas_algorithm.cc
@@ -19,15 +19,13 @@
  * for the compiler-research.org organization.
  */
 
-#include "diffusion_thomas_algorithm.h"
-#include "hyperparams.h"
+#include "diffusion/diffusion_thomas_algorithm.h"
 #include "core/agent/agent.h"
 #include "core/container/math_array.h"
 #include "core/diffusion/diffusion_grid.h"
 #include "core/param/param.h"
 #include "core/real_t.h"
 #include "core/resource_manager.h"
-#include "core/substance_interactor.h"
 #include <array>
 #include <cassert>
 #include <cstddef>
@@ -35,6 +33,8 @@
 #include <string>
 #include <utility>
 #include <vector>
+#include "../interfaces/substance_interactor.h"
+#include "params/hyperparams.h"
 
 namespace bdm {
