+++ src/diffusion/diffusion_thomas_algorithm.h
@@ -24,11 +24,11 @@
 
 #include "core/diffusion/diffusion_grid.h"
 #include "core/real_t.h"
-#include "core/substance_interactor.h"
 #include "core/util/root.h"
 #include <cstddef>
 #include <string>
 #include <vector>
+#include "../interfaces/substance_interactor.h"
 
 namespace bdm {
