+++ src/utils/utils_aux.cc
@@ -19,10 +19,7 @@
  * for the compiler-research.org organization.
  */
 
-#include "utils_aux.h"
-#include "cart_cell.h"
-#include "hyperparams.h"
-#include "tumor_cell.h"
+#include "utils/utils_aux.h"
 #include "core/agent/agent.h"
 #include "core/container/math_array.h"
 #include "core/diffusion/diffusion_grid.h"
@@ -41,6 +38,9 @@
 #include <memory>
 #include <tuple>
 #include <vector>
+#include "agents/cart_cell.h"
+#include "agents/tumor_cell.h"
+#include "params/hyperparams.h"
 
 namespace bdm {
